requirements.txt
pdfplumber
pytesseract
pillow
chromadb
sentence-transformers
python-dotenv
streamlit
pymupdf
